[FIX] Fixes wrong input for VoxelData constructor when saving Cerebellum direction vectors
This bug has been reported by Jonathan Lurie and Nabil Alibou (DKE Team):

The script generating the direction vectors of the mouse Cerebellum
crashes at the moment of saving the direction vectors array to disk.

The issue is solved by removing the extraneous (and erroneous)
VoxelData object around the array that is returned by the internal
method.

Besides, the other following fix has been done: the Isocortex direction
vectors are now saved to disk with the type float32.

There were supposed to be saved that way but the type specification was
actually missing. (Thanks to Jonathan for reporting this!)

+++ atlas_building_tools/direction_vectors/isocortex.py
@@ -86,7 +86,7 @@ def compute_direction_vectors(
         Vector field of 3D unit vectors over the isocortex volume with the same shape
         as the input one. Voxels outside the Isocortex have np.nan coordinates.
     '''
-    direction_vectors = np.full(brain_regions.shape + (3,), np.nan)
+    direction_vectors = np.full(brain_regions.shape + (3,), np.nan, dtype=np.float32)
     region_map = load_region_map(region_map)
     # Get the highest-level regions of the isocortex: ACAd, ACAv, AId, AIp, AIv, ...
     # In the AIBS mouse ccfv3 annotation, there are 43 isocortical regions.
